requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.13
aiosignal==1.3.2
attrs==25.1.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
discord.py==2.5.0
frozenlist==1.5.0
future==1.0.0
idna==3.10
multidict==6.1.0
propcache==0.3.0
pycparser==2.22
PyNaCl==1.5.0
python-dotenv==1.0.1
requests==2.32.3
urllib3==2.3.0
yarl==1.18.3